extract all http library code from executable
