+++ bind_socket.c
@@ -1,3 +1,8 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
+
 #include "bind_socket.h"
 
 /**
