+++ create_socket.h
@@ -1,7 +1,6 @@
 #ifndef CREATE_SOCKET_H
 #define CREATE_SOCKET_H
 
-#include <stdlib.h>
 #include <netinet/in.h>
 
 //Create a socket and assign current_socket to the descriptor
