+++ daemonize.c
@@ -1,5 +1,10 @@
 #include <pthread.h>
 #include <stdlib.h>
+#include <unistd.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+
+#include "daemonize.h"
 
 void daemonize(void)
 {
